requirements.txt
torch>=2.0
torch-geometric>=2.4
torch-scatter
torch-sparse
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
pyyaml>=6.0
requests>=2.31
tqdm>=4.65
matplotlib>=3.7
seaborn>=0.12
